feat: Modify Folder Structure
이벤트 구조 수정
